[REM] pos_(restaurant_)adyen: remove pos_restaurant_adyen module
Task: [#4841174](https://www.odoo.com/odoo/1737/tasks/4841174)

---

The `pos_restaurant_adyen` module is no longer needed since all its code is related to `set_tip_after_payment` feature,
which are now handeled by the `point_of_sale` module. This means that the code from `pos_restaurant_adyen`
has to be merged into `pos_adyen`.

closes odoo/odoo#240733

Related: odoo/upgrade#9119
