+++ addons/pos_adyen/controllers/__init__.py
@@ -1,2 +1,2 @@
-# coding: utf-8
 from . import main
+from . import payment_adyen
