+++ addons/pos_adyen/models/__init__.py
@@ -3,4 +3,5 @@
 from . import pos_config
 from . import pos_order
 from . import pos_payment_method
+from . import pos_payment
 from . import res_config_settings
