+++ addons/pos_adyen/models/pos_payment.py
@@ -1,4 +1,3 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
 from odoo import models
 
 TIMEOUT = 10
