+++ addons/pos_adyen/static/tests/unit/data/pos_payment_method.data.js
@@ -3,6 +3,10 @@ import { PosPaymentMethod } from "@point_of_sale/../tests/unit/data/pos_payment_
 
 patch(PosPaymentMethod.prototype, {
     _load_pos_data_fields() {
-        return [...super._load_pos_data_fields(), "adyen_terminal_identifier"];
+        return [
+            ...super._load_pos_data_fields(),
+            "adyen_terminal_identifier",
+            "adyen_merchant_account",
+        ];
     },
 });
