--- addons/pos_restaurant_adyen/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import controllers
-from . import models
