--- addons/pos_restaurant_adyen/__manifest__.py
@@ -1,23 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-{
-    'name': 'POS Restaurant Adyen',
-    'category': 'Point of Sale',
-    'sequence': 6,
-    'summary': 'Adds American style tipping to Adyen',
-    'depends': ['pos_adyen', 'pos_restaurant', 'payment_adyen'],
-    'data': [
-        'views/pos_payment_method_views.xml',
-        ],
-    'auto_install': True,
-    'assets': {
-        'point_of_sale._assets_pos': [
-            'pos_restaurant_adyen/static/src/**/*',
-        ],
-        'web.assets_unit_tests': [
-            'pos_restaurant_adyen/static/tests/unit/data/**/*'
-        ],
-    },
-    'author': 'Odoo S.A.',
-    'license': 'LGPL-3',
-}
