--- addons/pos_restaurant_adyen/controllers/main.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-import logging
-
-from odoo import http
-from odoo.http import request
-from odoo.addons.payment_adyen.controllers.main import AdyenController
-
-_logger = logging.getLogger(__name__)
-
-
-class PosRestaurantAdyenController(AdyenController):
-
-    @http.route()
-    def adyen_webhook(self, **post):
-        if post.get('eventCode') in ['CAPTURE', 'AUTHORISATION_ADJUSTMENT'] and post.get('success') != 'true':
-                _logger.warning('%s for transaction_id %s failed', post.get('eventCode'), post.get('originalReference'))
-        return super(PosRestaurantAdyenController, self).adyen_webhook(**post)
