--- addons/pos_restaurant_adyen/i18n/pos_restaurant_adyen.pot
@@ -1,55 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* pos_restaurant_adyen
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 19.1a1\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-12-16 18:36+0000\n"
-"PO-Revision-Date: 2025-12-16 18:36+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__adyen_merchant_account
-msgid "Adyen Merchant Account"
-msgstr ""
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_order__display_name
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment__display_name
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_order__id
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment__id
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__id
-msgid "ID"
-msgstr ""
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_order
-msgid "Point of Sale Order"
-msgstr ""
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_payment
-msgid "Point of Sale Payment"
-msgstr ""
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_payment_method
-msgid "Point of Sale Payment Method"
-msgstr ""
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,help:pos_restaurant_adyen.field_pos_payment_method__adyen_merchant_account
-msgid "The POS merchant account code used in Adyen"
-msgstr ""
