--- addons/pos_restaurant_adyen/i18n/ru.po
@@ -1,48 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* pos_restaurant_adyen
-#
-# Translators:
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 17.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-10-26 21:55+0000\n"
-"PO-Revision-Date: 2024-01-30 15:14+0400\n"
-"Last-Translator: \n"
-"Language-Team: Russian (https://app.transifex.com/odoo/teams/41243/ru/)\n"
-"Language: ru\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);\n"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__adyen_merchant_account
-msgid "Adyen Merchant Account"
-msgstr "Торговый счет Adyen"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_order
-msgid "Point of Sale Orders"
-msgstr "Заказы в торговых точках"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_payment_method
-msgid "Point of Sale Payment Methods"
-msgstr "Способы оплаты в торговых точках"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_payment
-msgid "Point of Sale Payments"
-msgstr "Платежи в точках продаж"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_session
-msgid "Point of Sale Session"
-msgstr "Сессия в торговой точке"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,help:pos_restaurant_adyen.field_pos_payment_method__adyen_merchant_account
-msgid "The POS merchant account code used in Adyen"
-msgstr "Код торгового счета POS, используемый в Adyen"
