--- addons/pos_restaurant_adyen/i18n/zh_TW.po
@@ -1,59 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* pos_restaurant_adyen
-# 
-# Translators:
-# Wil Odoo, 2025
-# 
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~18.2\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-02-20 10:02+0000\n"
-"PO-Revision-Date: 2025-05-06 12:04+0000\n"
-"Last-Translator: Wil Odoo, 2025\n"
-"Language-Team: Chinese (Taiwan) (https://app.transifex.com/odoo/teams/41243/zh_TW/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: zh_TW\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__adyen_merchant_account
-msgid "Adyen Merchant Account"
-msgstr "Adyen 商家賬號"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_order__display_name
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment__display_name
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__display_name
-msgid "Display Name"
-msgstr "顯示名稱"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_order__id
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment__id
-#: model:ir.model.fields,field_description:pos_restaurant_adyen.field_pos_payment_method__id
-msgid "ID"
-msgstr "識別號"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_order
-msgid "Point of Sale Orders"
-msgstr "POS訂單"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_payment_method
-msgid "Point of Sale Payment Methods"
-msgstr "POS付款條件"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model,name:pos_restaurant_adyen.model_pos_payment
-msgid "Point of Sale Payments"
-msgstr "POS付款"
-
-#. module: pos_restaurant_adyen
-#: model:ir.model.fields,help:pos_restaurant_adyen.field_pos_payment_method__adyen_merchant_account
-msgid "The POS merchant account code used in Adyen"
-msgstr "Adyen 中使用的 POS 商戶帳號代碼"
