--- addons/pos_restaurant_adyen/models/__init__.py
@@ -1,4 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import pos_payment
-from . import pos_payment_method
