--- addons/pos_restaurant_adyen/static/src/overrides/models/payment_adyen.js
@@ -1,18 +1,0 @@
-import { PaymentAdyen } from "@pos_adyen/app/utils/payment/payment_adyen";
-import { patch } from "@web/core/utils/patch";
-
-patch(PaymentAdyen.prototype, {
-    _adyenPayData() {
-        var data = super._adyenPayData(...arguments);
-
-        if (data.SaleToPOIRequest.PaymentRequest.SaleData.SaleToAcquirerData) {
-            data.SaleToPOIRequest.PaymentRequest.SaleData.SaleToAcquirerData +=
-                "&authorisationType=PreAuth";
-        } else {
-            data.SaleToPOIRequest.PaymentRequest.SaleData.SaleToAcquirerData =
-                "authorisationType=PreAuth";
-        }
-
-        return data;
-    },
-});
