--- addons/pos_restaurant_adyen/static/tests/unit/data/pos_payment_method.data.js
@@ -1,8 +1,0 @@
-import { patch } from "@web/core/utils/patch";
-import { PosPaymentMethod } from "@point_of_sale/../tests/unit/data/pos_payment_method.data";
-
-patch(PosPaymentMethod.prototype, {
-    _load_pos_data_fields() {
-        return [...super._load_pos_data_fields(), "adyen_merchant_account"];
-    },
-});
